requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Login==0.6.3
psycopg[binary]==3.2.6
python-dotenv==1.0.1
gunicorn==23.0.0
Pillow==11.1.0
